+++ archiver/biostudies.py
@@ -1,9 +1,8 @@
-from archiver.dsp_post_process import dsp_attribute, fixed_dsp_attribute
+from archiver.dsp_post_process import dsp_attribute
 from conversion.json_mapper import JsonMapper
 from conversion.post_process import *
 
 PREFIX_PROJECT = 'project_'
-PREFIX_STUDY = 'study_'
 
 
 def _parse_name(*args):
@@ -61,24 +60,3 @@ project_spec = {
 
 def convert_project(hca_data: dict):
     return JsonMapper(hca_data).map(project_spec)
-
-
-study_spec = {
-    '$on': 'project',
-    'alias': ['uuid.uuid', prefix_with, (PREFIX_STUDY)],
-    'attributes': {
-        'HCA Project UUID': ['uuid.uuid', dsp_attribute],
-        'Project Core - Project Short Name': ['content.project_core.project_short_name', dsp_attribute],
-        'study_type': ['', fixed_dsp_attribute, 'Transcriptome Analysis'],
-        'study_abstract': ['content.project_core.project_description', dsp_attribute],
-    },
-    'title': ['content.project_core.project_title'],
-    'description': ['content.project_core.project_description'],
-    'projectRef': {
-        'alias': ['', default_to, '{projectAlias.placeholder}']
-    }
-}
-
-
-def convert_study(hca_data: dict):
-    return JsonMapper(hca_data).map(study_spec)
